kindle/requirements.txt
lxml == 3.5.0
requests == 2.9.1
AmazonAPIWrapper == 0.0.11
beautifulsoup4 == 4.5.1
